# backend/swagger.yaml
openapi: 3.0.0
info:
  title: SalesBase CRM API
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Healthy status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                  version:
                    type: string
        "500":
          description: Internal server error
  /api/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
        "401":
          description: Unauthorized
        "400":
          description: Bad request
  /api/auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: User registered
        "400":
          description: Bad request
  /api/auth/logout:
    post:
      summary: Logout
      responses:
        "200":
          description: Logged out
  /api/contacts:
    get:
      summary: List contacts
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Contact"
        "401":
          description: Unauthorized
    post:
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contact"
      responses:
        "201":
          description: Contact created
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /api/contacts/{id}:
    get:
      summary: Get contact by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
        "404":
          description: Contact not found
        "401":
          description: Unauthorized
    put:
      summary: Update contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Contact"
      responses:
        "200":
          description: Contact updated
        "400":
          description: Bad request
        "404":
          description: Contact not found
        "401":
          description: Unauthorized
    delete:
      summary: Delete contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Contact deleted
        "404":
          description: Contact not found
        "401":
          description: Unauthorized
  /api/companies:
    get:
      summary: List companies
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: object
                properties:
                  companies:
                    type: array
                    items:
                      $ref: "#/components/schemas/Company"
        "401":
          description: Unauthorized
  /api/companies/{id}:
    get:
      summary: Get company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
        "404":
          description: Company not found
        "401":
          description: Unauthorized
    put:
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Company"
      responses:
        "200":
          description: Company updated
        "400":
          description: Bad request
        "404":
          description: Company not found
        "401":
          description: Unauthorized
    delete:
      summary: Delete company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Company deleted
        "404":
          description: Company not found
        "401":
          description: Unauthorized
  /api/deals:
    get:
      summary: List deals
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of deals
          content:
            application/json:
              schema:
                type: object
                properties:
                  deals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Deal"
        "401":
          description: Unauthorized
    post:
      summary: Create deal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Deal"
      responses:
        "201":
          description: Deal created
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /api/deals/{id}:
    get:
      summary: Get deal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deal details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deal"
        "404":
          description: Deal not found
        "401":
          description: Unauthorized
    put:
      summary: Update deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Deal"
      responses:
        "200":
          description: Deal updated
        "400":
          description: Bad request
        "404":
          description: Deal not found
        "401":
          description: Unauthorized
    delete:
      summary: Delete deal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deal deleted
        "404":
          description: Deal not found
        "401":
          description: Unauthorized
  /api/pipeline/analytics/overview:
    get:
      summary: Pipeline analytics overview
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  pipeline_stages:
                    type: array
                    items:
                      $ref: "#/components/schemas/PipelineStageAnalytics"
                  pipeline_summary:
                    $ref: "#/components/schemas/PipelineSummary"
                  last_updated:
                    type: string
        "401":
          description: Unauthorized
  /api/reports:
    get:
      summary: List reports
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      type: object
        "401":
          description: Unauthorized
  /api/reports/export:
    get:
      summary: Export report as CSV
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Unauthorized
  /api/search:
    get:
      summary: Search contacts, companies, deals
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
        "401":
          description: Unauthorized
  /api/integrations/zapier:
    post:
      summary: Zapier webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "400":
          description: Bad request
  /api/integrations/slack:
    post:
      summary: Slack webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "400":
          description: Bad request
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Contact:
      type: object
      required:
        - id
        - first_name
        - last_name
        - email
      properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
    Company:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
        name:
          type: string
        industry:
    Deal:
      type: object
      required:
        - id
        - company_id
        - value
        - status
        - pipeline_stage_id
        - created_at
      properties:
        id:
          type: integer
        company_id:
          type: integer
        value:
          type: number
        status:
          type: string
        pipeline_stage_id:
          type: integer
        created_at:
          type: string
        value:
          type: number
        status:
          type: string
        pipeline_stage_id:
          type: integer
        created_at:
          type: string
    PipelineStageAnalytics:
      type: object
      properties:
        stage_name:
          type: string
        display_order:
          type: integer
        win_probability:
          type: number
        deal_count:
          type: integer
        total_value:
          type: number
        avg_deal_value:
          type: number
        avg_probability:
          type: number
        deals_last_30_days:
          type: integer
        won_deals:
          type: integer
        lost_deals:
          type: integer
        win_rate:
          type: number
        conversion_rate:
          type: number
    PipelineSummary:
      type: object
      properties:
        total_deals:
          type: integer
        total_value:
          type: number
        avg_win_rate:
          type: number
        avg_days_in_pipeline:
          type: number
        new_deals_30_days:
          type: integer
        won_deals_30_days:
          type: integer
        lost_deals_30_days:
          type: integer
security:
  - bearerAuth: []